Split Main.py into several Classes and Files
