+++ .gitignore
@@ -1,0 +1,5 @@
+__pycache__/constants.cpython-311.pyc
+__pycache__/DataProcessing.cpython-311.pyc
+__pycache__/GraphPlot.cpython-311.pyc
+__pycache__/InteractiveInterface.cpython-311.pyc
+__pycache__/TimeConversions.cpython-311.pyc
